feat: custom plugins may be passed
