+++ packages/verified-fetch/src/errors.ts
@@ -24,38 +24,3 @@ export class SubdomainNotSupportedError extends Error {
     this.name = 'SubdomainNotSupportedError'
   }
 }
-
-export interface PluginErrorOptions {
-  fatal?: boolean
-  details?: Record<string, any>
-  response?: Response
-}
-
-export interface FatalPluginErrorOptions extends PluginErrorOptions {
-  response: Response
-}
-
-export class PluginError extends Error {
-  public name = 'PluginError'
-  public code: string
-  public fatal: boolean
-  public details?: Record<string, any>
-  public response?: any
-
-  constructor (code: string, message: string, options?: PluginErrorOptions) {
-    super(message)
-    this.code = code
-    this.fatal = options?.fatal ?? false
-    this.details = options?.details
-    this.response = options?.response
-  }
-}
-
-export class PluginFatalError extends PluginError {
-  public name = 'PluginFatalError'
-
-  constructor (code: string, message: string, options: FatalPluginErrorOptions) {
-    super(code, message, { ...options, fatal: true })
-    this.name = 'PluginFatalError'
-  }
-}
