+++ packages/verified-fetch/src/plugins/index.ts
@@ -1,0 +1,7 @@
+/**
+ * This file is the entry into all things we export from the `src/plugins` directory.
+ */
+
+export { PluginError, PluginFatalError } from './errors.js'
+export { BasePlugin } from './plugin-base.js'
+export type { PluginOptions, PluginContext } from './types.js'
