+++ packages/verified-fetch/src/plugins/types.ts
@@ -1,4 +1,4 @@
-import type { PluginError } from '../errors.js'
+import type { PluginError } from './errors.js'
 import type { VerifiedFetchInit } from '../index.js'
 import type { ContentTypeParser, RequestFormatShorthand } from '../types.js'
 import type { ParsedUrlStringResults } from '../utils/parse-url-string.js'
@@ -11,8 +11,7 @@ import type { CID } from 'multiformats/cid'
 import type { CustomProgressEvent } from 'progress-events'
 
 /**
- * Contains all configuration needed to handle a request.
- * - User Config: signal, timeouts, onProgress callbacks, booleans like withServerTiming, etc.
+ * Contains common components and functions required by plugins to handle a request.
  * - Read-Only: Plugins can read but shouldn’t rewrite them.
  * - Persistent: Relevant even after the request completes (e.g., logging or metrics).
  */
@@ -40,7 +39,6 @@ export interface PluginContext {
    * A plugin should increment this value if it modifies the context.
    */
   modified: number
-  blockstore?: Blockstore
   withServerTiming?: boolean
   onProgress?(evt: CustomProgressEvent<any>): void
   options?: Omit<VerifiedFetchInit, 'signal'> & AbortOptions
@@ -54,9 +52,23 @@ export interface PluginContext {
   [key: string]: unknown
 }
 
-export interface FetchHandlerPlugin {
+export interface VerifiedFetchPlugin {
   readonly codes: number[]
   readonly log: Logger
   canHandle (context: PluginContext): boolean
   handle (context: PluginContext): Promise<Response | null>
+}
+
+export interface VerifiedFetchPluginFactory {
+  (options: PluginOptions): VerifiedFetchPlugin
+}
+
+export interface PluginErrorOptions {
+  fatal?: boolean
+  details?: Record<string, any>
+  response?: Response
+}
+
+export interface FatalPluginErrorOptions extends PluginErrorOptions {
+  response: Response
 }
